+++ CoordinateSystems/GeographicCoordinateSystem.cs
@@ -9,38 +9,40 @@ namespace TensionDev.CoordinateSystems
     /// </summary>
     public class GeographicCoordinateSystem
     {
-        private double latitude;
-        private double longitude;
-        private double altitude;
+        private const double FEET_TO_METRES = 0.3048;
+
+        private double _latitude;
+        private double _longitude;
+        private double _altitude;
 
         /// <summary>
         /// Latitude in Radians
         /// </summary>
-        public Double LatitudeDecimalRadians { get => latitude * Math.PI / 180.0; set => LatitudeDecimalDegrees = value * 180.0 / Math.PI; }
+        public Double LatitudeDecimalRadians { get => _latitude * Math.PI / 180.0; set => LatitudeDecimalDegrees = value * 180.0 / Math.PI; }
 
         /// <summary>
         /// Longitude in Radians
         /// </summary>
-        public Double LongitudeDecimalRadians { get => longitude * Math.PI / 180.0; set => LongitudeDecimalDegrees = value * 180.0 / Math.PI; }
+        public Double LongitudeDecimalRadians { get => _longitude * Math.PI / 180.0; set => LongitudeDecimalDegrees = value * 180.0 / Math.PI; }
 
         /// <summary>
         /// Altitude in Feet
         /// </summary>
-        public Double AltitudeFeet { get => altitude / 0.3048; set => altitude = value * 0.3048; }
+        public Double AltitudeFeet { get => _altitude / FEET_TO_METRES; set => _altitude = value * FEET_TO_METRES; }
 
         /// <summary>
         /// Latitude in Degrees
         /// </summary>
-        public Double LatitudeDecimalDegrees { get => latitude; set => latitude = Math.Max(Math.Min(value, 90), -90); }
+        public Double LatitudeDecimalDegrees { get => _latitude; set => _latitude = Math.Max(Math.Min(value, 90), -90); }
 
         /// <summary>
         /// Longitude in Degrees
         /// </summary>
-        public Double LongitudeDecimalDegrees { get => longitude; set => longitude = Math.Max(Math.Min(value, 180), -180); }
+        public Double LongitudeDecimalDegrees { get => _longitude; set => _longitude = Math.Max(Math.Min(value, 180), -180); }
 
         /// <summary>
         /// Altitude in Metres
         /// </summary>
-        public Double AltitudeMetres { get => altitude; set => altitude = value; }
+        public Double AltitudeMetres { get => _altitude; set => _altitude = value; }
     }
 }